Moved util.counters to internal.util.counters
