+++ hazelcast/src/main/java/com/hazelcast/internal/metrics/impl/MethodProbe.java
@@ -20,7 +20,7 @@ import com.hazelcast.internal.metrics.DoubleProbeFunction;
 import com.hazelcast.internal.metrics.LongProbeFunction;
 import com.hazelcast.internal.metrics.Probe;
 import com.hazelcast.internal.metrics.ProbeFunction;
-import com.hazelcast.util.counters.Counter;
+import com.hazelcast.internal.util.counters.Counter;
 
 import java.lang.reflect.Method;
 import java.util.Collection;
