+++ hazelcast/src/main/java/com/hazelcast/internal/util/counters/MwCounter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.util.counters;
+package com.hazelcast.internal.util.counters;
 
 import java.util.concurrent.atomic.AtomicLongFieldUpdater;
