+++ hazelcast/src/main/java/com/hazelcast/internal/util/counters/package-info.java
@@ -17,4 +17,4 @@
 /**
  * Provides various counter implementations.
  */
-package com.hazelcast.util.counters;
+package com.hazelcast.internal.util.counters;
