+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/SocketReader.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.nio.tcp;
 
-import com.hazelcast.util.counters.Counter;
+import com.hazelcast.internal.util.counters.Counter;
 
 /**
  * The SocketReader is responsible for reading data from the socket, on behalf of a connection, into a
