+++ hazelcast/src/test/java/com/hazelcast/internal/util/counters/SafeSwCounterTest.java
@@ -1,4 +1,4 @@
-package com.hazelcast.util.counters;
+package com.hazelcast.internal.util.counters;
 
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
@@ -7,17 +7,16 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import static com.hazelcast.util.counters.MwCounter.newMwCounter;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(HazelcastSerialClassRunner.class)
 @Category(QuickTest.class)
-public class MwCounterTest {
-    private MwCounter counter;
+public class SafeSwCounterTest {
+    private SwCounter.SafeSwCounter counter;
 
     @Before
     public void setup() {
-        counter = newMwCounter();
+        counter = new SwCounter.SafeSwCounter(0);
     }
 
     @Test
