+++ hazelcast/src/test/java/com/hazelcast/internal/util/counters/SwCounterTest.java
@@ -1,4 +1,4 @@
-package com.hazelcast.util.counters;
+package com.hazelcast.internal.util.counters;
 
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
@@ -7,31 +7,34 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
+import static com.hazelcast.internal.util.counters.SwCounter.newSwCounter;
 import static org.junit.Assert.assertEquals;
 
+/**
+ * If Unsafe is available, this will trigger the testing of the {@link SwCounter.UnsafeSwCounter} class. Otherwise
+ * the {@link SwCounter.SafeSwCounter} is tested.
+ */
 @RunWith(HazelcastSerialClassRunner.class)
 @Category(QuickTest.class)
-public class SafeSwCounterTest {
-    private SwCounter.SafeSwCounter counter;
+public class SwCounterTest {
+
+    private SwCounter counter;
 
     @Before
     public void setup() {
-        counter = new SwCounter.SafeSwCounter(0);
+        counter = newSwCounter();
     }
 
     @Test
     public void inc() {
-        counter.inc();
-        assertEquals(1, counter.get());
+        assertEquals(1, counter.inc());
     }
 
     @Test
     public void inc_withAmount() {
-        counter.inc(10);
-        assertEquals(10, counter.get());
-
-        counter.inc(0);
-        assertEquals(10, counter.get());
+        assertEquals(10, counter.inc(10));
+        assertEquals(10, counter.inc(0));
+        assertEquals(0, counter.inc(-10));
     }
 
     @Test
